refactor: shrink library surface and fix latest subcommand alias
Keep sem-tool as a CLI crate with private modules and a hidden test_support
module for integration tests. Use clap visible_alias for latest on max and
replace rustdoc markdown emphasis in help docs with plain text.
